+++ src/rer/customersatisfaction/browser/adapters.py
@@ -21,7 +21,7 @@ class CustomerSatisfactionBodyClasses(object):
         context_state = api.content.get_view(
             context=self.context,
             request=self.request,
-            name=u"plone_context_state",
+            name="plone_context_state",
         )
         if context_state.canonical_object() == api.portal.get():
             return []