+++ src/rer/customersatisfaction/restapi/services/sites_list/get.py
@@ -8,13 +8,15 @@ class View(Overview):
         data = []
         for site in self.sites():
             site_url = site.portal_url()
-            data.append(
-                {
-                    "id": site.id,
-                    "url": site_url,
-                    "title": site.title,
-                }
-            )
+
+            if site.id not in [i.get("id", None) for i in data]:
+                data.append(
+                    {
+                        "id": site.id,
+                        "url": site_url,
+                        "title": site.title,
+                    }
+                )
 
         self.set_headers()