+++ src/rer/customersatisfaction/storage/catalog.py
@@ -13,11 +13,11 @@ class CustomerSatisfactionSoupCatalogFactory(object):
     def __call__(self, context):
         catalog = Catalog()
         text_indexer = NodeTextIndexer(["title"])
-        catalog[u"text"] = CatalogTextIndex(text_indexer)
+        catalog["text"] = CatalogTextIndex(text_indexer)
         vote_indexer = NodeAttributeIndexer("vote")
-        catalog[u"vote"] = CatalogFieldIndex(vote_indexer)
+        catalog["vote"] = CatalogFieldIndex(vote_indexer)
         uid_indexer = NodeAttributeIndexer("uid")
-        catalog[u"uid"] = CatalogFieldIndex(uid_indexer)
+        catalog["uid"] = CatalogFieldIndex(uid_indexer)
         date_indexer = NodeAttributeIndexer("date")
-        catalog[u"date"] = CatalogFieldIndex(date_indexer)
+        catalog["date"] = CatalogFieldIndex(date_indexer)
         return catalog