+++ src/rer/customersatisfaction/tests/test_show_feedbacks_view.py
@@ -16,7 +16,6 @@ import unittest
 
 
 class TestShowFeedbacksTool(unittest.TestCase):
-
     layer = RER_CUSTOMERSATISFACTION_FUNCTIONAL_TESTING
 
     def setUp(self):