+++ CHANGES.rst
@@ -22,9 +22,10 @@ Changelog
 
 2.2.1 (2023-03-24)
 ------------------
-
+- Add export to customer-satisfaction-global view
+  [folix-01]
 - Add customer-satisfaction-global view
-  [foxtrot-dfm1]
+  [folix-01]
 - Fix english label.
   [cekk]